requirements.txt
# Core
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# PDF extraction
PyMuPDF==1.25.5
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.2.1

# Chunking
langchain-text-splitters==0.3.8

# Vector store
chromadb==0.6.3

# LLM client
httpx==0.28.1

# Database
aiosqlite==0.21.0

# Export
fpdf2==2.8.3

# Frontend
streamlit==1.44.1
requests==2.32.3
